+++ src/index.html
@@ -28,7 +28,7 @@
       </form>    
     </div>
 
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 
 </body>
 </html>
